refactor(errors): separate provider localization
