+++ apps/web/lib/errors/provider-unavailable-message.ts
@@ -2,9 +2,9 @@ import 'server-only';
 
 import type { ProviderId } from '@/lib/provider-errors';
 import { getProviderUnavailableDetails } from '@/lib/provider-errors';
-import messages from '@/messages/en.json';
 
-const PROVIDER_UNAVAILABLE_TEMPLATE = messages.errorCodes.PROVIDER_UNAVAILABLE;
+export const PROVIDER_UNAVAILABLE_TEMPLATE =
+  '{provider} is temporarily unavailable. Please retry.';
 
 export function getProviderUnavailableMessage(provider: ProviderId): string {
   const { provider: displayName } = getProviderUnavailableDetails(provider);
