+++ apps/web/messages/it.json
@@ -708,7 +708,6 @@
       "guardrailViolation": "Questa richiesta è stata bloccata da una policy di sicurezza di terze parti per la clonazione vocale. Prova con un testo diverso o con un altro audio di riferimento.",
       "insufficientCredits": "Servono __CREDITS__ crediti per clonare questo audio.",
       "internalError": "Impossibile clonare la voce. Riprova.",
-      "providerUnavailable": "Il provider di clonazione vocale è temporaneamente non disponibile. Riprova.",
       "invalidContentType": "La richiesta di caricamento non è valida. Riprova.",
       "invalidFileType": "Tipo di file non valido. Carica MP3, OGG, Opus, M4A, WAV o WebM.",
       "missingLocale": "Seleziona una lingua.",
@@ -1200,6 +1199,6 @@
   "errorCodes": {
     "VALIDATION_ERROR_DISPOSABLE_EMAIL": "Gli indirizzi email temporanei non sono consentiti",
     "AUTH_PROVIDER_RATELIMIT": "Per motivi di sicurezza, puoi richiedere questo solo dopo {time} secondi.",
-    "PROVIDER_UNAVAILABLE": "{provider} non è temporaneamente disponibile. Riprova."
+    "PROVIDER_UNAVAILABLE": "{provider} non è disponibile al momento. Riprova."
   }
 }
